+++ src/main/java/com/paiondata/aristotle/mapper/impl/GraphMapperImpl.java
@@ -19,11 +19,18 @@ import com.paiondata.aristotle.common.base.Constants;
 import com.paiondata.aristotle.common.util.Neo4jUtil;
 import com.paiondata.aristotle.mapper.GraphMapper;
 import com.paiondata.aristotle.model.entity.Graph;
+
+import org.neo4j.driver.Driver;
 import org.neo4j.driver.Record;
+import org.neo4j.driver.Session;
+import org.neo4j.driver.SessionConfig;
 import org.neo4j.driver.Transaction;
 import org.neo4j.driver.Values;
 import org.springframework.stereotype.Repository;
 
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -31,6 +38,16 @@ import java.util.Map;
  */
 @Repository
 public class GraphMapperImpl implements GraphMapper {
+
+    private final Driver driver;
+
+    /**
+     * Constructs a GraphMapperImpl object with the specified Driver.
+     * @param driver the Driver instance
+     */
+    public GraphMapperImpl(Driver driver) {
+        this.driver = driver;
+    }
 
     /**
      * Creates a new graph with the specified properties.
@@ -75,5 +92,66 @@ public class GraphMapperImpl implements GraphMapper {
                 .createTime((String) objectMap.get(Constants.CREATE_TIME))
                 .updateTime((String) objectMap.get(Constants.UPDATE_TIME))
                 .build();
+    }
+
+    /**
+     * Retrieves users' associated graphs by UIDCID.
+     *
+     * @param uidcid the UIDCID of the user
+     *
+     * @return a list of maps containing user information and associated graphs
+     */
+    @Override
+    public List<Map<String, Object>> getGraphsByUidcid(final String uidcid) {
+        final String cypherQuery = "MATCH (u:User)-[r:RELATION]->(g:Graph) WHERE u.uidcid = $uidcid RETURN DISTINCT g";
+
+        try (Session session = driver.session(SessionConfig.builder().build())) {
+            return session.readTransaction(tx -> {
+                final var result = tx.run(cypherQuery, Values.parameters("uidcid", uidcid));
+                final List<Map<String, Object>> resultList = new ArrayList<>();
+                while (result.hasNext()) {
+                    final Record record = result.next();
+                    final Map<String, Object> graphNode = Neo4jUtil.extractGraph(record.get("g"));
+
+                    resultList.add(graphNode);
+                }
+                return resultList;
+            });
+        }
+    }
+
+    /**
+     * Updates a graph by its UUID.
+     *
+     * @param uuid the UUID of the graph
+     * @param title the new title of the graph
+     * @param description the new description of the graph
+     * @param currentTime the current time for update
+     * @param tx the Neo4j transaction
+     */
+
+    @Override
+    public void updateGraphByUuid(final String uuid, final String title,
+                                  final String description, final String currentTime, final Transaction tx) {
+        final StringBuilder cypherQuery = new StringBuilder("MATCH (g:Graph { uuid: $uuid }) ");
+        final Map<String, Object> parameters = new HashMap<>();
+        parameters.put(Constants.UUID, uuid);
+
+        if (title != null) {
+            cypherQuery.append("SET g.title = $title ");
+            parameters.put(Constants.TITLE, title);
+        }
+
+        if (description != null) {
+            cypherQuery.append("SET g.description = $description ");
+            parameters.put(Constants.DESCRIPTION, description);
+        }
+
+        cypherQuery.append("SET g.update_time = $updateTime ");
+        parameters.put(Constants.UPDATE_TIME, currentTime);
+
+        cypherQuery.append("RETURN g");
+
+        tx.run(cypherQuery.toString(), parameters);
     }
 }
