+++ src/main/java/com/paiondata/aristotle/repository/GraphRepository.java
@@ -42,31 +42,6 @@ public interface GraphRepository extends Neo4jRepository<Graph, Long> {
     Graph getGraphByUuid(String uuid);
 
     /**
-     * Creates and binds a new graph to an existing user.
-     *
-     * @param title           the title of the graph
-     * @param description     the description of the graph
-     * @param userUidcid      the UID/CID of the user
-     * @param graphUuid       the UUID of the new graph
-     * @param relationUuid    the UUID of the relationship
-     * @param currentTime     the current timestamp
-     * @return the created graph
-     */
-    @Query("MATCH (u:User) WHERE u.uidcid = $userUidcid "
-            + "CREATE (g:Graph {uuid: $graphUuid, title: $title, description: $description, "
-            + "create_time: $currentTime, update_time: $currentTime}) "
-            + "WITH u, g "
-            + "CREATE (u)-[r:RELATION {name: 'HAVE', uuid: $relationUuid, create_time: $currentTime, "
-            + "update_time: $currentTime}]->(g) "
-            + "RETURN g")
-    Graph createAndBindGraph(@Param("title") String title,
-                             @Param("description") String description,
-                             @Param("userUidcid") String userUidcid,
-                             @Param("graphUuid") String graphUuid,
-                             @Param("relationUuid") String relationUuid,
-                             @Param("currentTime") String currentTime);
-
-    /**
      * Deletes graphs by their UUIDs.
      *
      * @param uuids the list of UUIDs of the graphs to be deleted
