+++ src/main/java/com/paiondata/aristotle/service/GraphService.java
@@ -15,15 +15,11 @@
  */
 package com.paiondata.aristotle.service;
 
-import com.paiondata.aristotle.model.dto.GraphCreateDTO;
 import com.paiondata.aristotle.model.dto.GraphDeleteDTO;
 import com.paiondata.aristotle.model.dto.GraphUpdateDTO;
-import com.paiondata.aristotle.model.entity.Graph;
 import com.paiondata.aristotle.model.vo.GraphVO;
 
 import org.neo4j.driver.Transaction;
-
-import java.util.Optional;
 
 /**
  * Service implementation for managing graphs.
@@ -41,25 +37,6 @@ public interface GraphService {
     GraphVO getGraphVOByUuid(String uuid);
 
     /**
-     * Retrieves a graph by its UUID.
-     *
-     * @param uuid the UUID of the graph
-     *
-     * @return an optional containing the graph if found, or empty if not found
-     */
-    Optional<Graph> getGraphByUuid(String uuid);
-
-    /**
-     * Creates and binds a new graph using the provided DTO.
-     *
-     * @param graphCreateDTO the DTO containing details to create a new graph
-     * @param tx the Neo4j transaction
-     *
-     * @return the created graph
-     */
-    Graph createAndBindGraph(GraphCreateDTO graphCreateDTO, Transaction tx);
-
-    /**
      * Deletes graphs by their UUIDs.
      *
      * @param graphDeleteDTO the DTO containing the UUIDs of the graphs to delete
@@ -70,6 +47,7 @@ public interface GraphService {
      * Updates a graph using the provided DTO.
      *
      * @param graphUpdateDTO the DTO containing details to update an existing graph
+     * @param tx the Neo4j transaction
      */
-    void updateGraph(GraphUpdateDTO graphUpdateDTO);
+    void updateGraph(GraphUpdateDTO graphUpdateDTO, Transaction tx);
 }
